pyproject.toml
[project]
name = "django-inertia-react"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.4",
    "django-ninja>=1.4.3",
    "django-vite>=3.1.0",
    "inertia-django>=1.2.0",
    "psycopg>=3.2.9",
]
